Add support for KMZ
Google Maps Engine starts returning KMZ instead of KML. This adds
support for unzipping the KMZ before parsing.

Because extraction of the KMZ works on stream, the `res` argument
in `request`'s callback, whose data are already consumed, cannot
be piped to `unzip.Parse`. Therefore, we listen to `response`
event of `request`, which gives us unmodified http.IncomingMessage
object that we can pipe.

This, however, does not give us the response's body, so we have to
do the crazy thing of listen to `data` and `end` events.

Other changes are just white spaces.